+++ src/types/post.ts
@@ -1,6 +1,11 @@
+import { User } from './user';
+import { Comment } from './comment';
+
 export interface Post {
-  userId: number,
+  userId: number;
   id: number,
-  title: string,
-  body: string,
+  title: string;
+  body: string;
+  user: User | null,
+  comments: Comment[],
 }
